♻️🏢 Extract route loader
